stellar_horizon::api
Module aggregations

Structs

AllTradesRequest

Request aggregated trade data.

FeeStatsRequest

Request fee stats.

OrderBookRequest

Request order book data.

PathsStrictReceiveRequest

Request paths for path payment strict receive.

PathsStrictSendRequest

Request paths for path payment strict send.

Enums

Resolution

Trade aggregation resolution.

Functions

all_trades

Creates a request to retrieve aggregated trade data.

fee_stats

Creates a request to retrieve statistics about network fees.

order_book

Creates a request to retrieve order book data.

paths_strict_receive

Creates a request to retrieve information about potential path payments.

paths_strict_send

Creates a request to retrieve information about potential path payments.