stellar_horizon::api::aggregations
Function all_trades

pub fn all_trades(
    start_time: DateTime<Utc>,
    end_time: DateTime<Utc>,
    resolution: Resolution,
    base_asset: Asset,
    counter_asset: Asset,
) -> AllTradesRequest

Creates a request to retrieve aggregated trade data.

A trade aggregation represents aggregated statistics on an asset pair (base and counter) for a specific time period. Trade aggregations are useful to developers of trading clients and provide historical trade data.

This endpoint displays trade data based on filters set in the arguments.

This is done by dividing a given time range into segments and aggregating statistics, for a given asset pair (base, counter) over each of these segments.

The duration of the segments is specified with the resolution parameter. The start and end of the time range are given by start_time and end_time respectively, which are both rounded to the nearest multiple of resolution since epoch.

The individual segments are also aligned with multiples of resolution since epoch. If you want to change this alignment, the segments can be offset by specifying the offset parameter.