stellar_horizon::api::aggregations
Function paths_strict_send

pub fn paths_strict_send<S: TryInto<Stroops>>(
    source_asset: Asset,
    destination_assets: Vec<CreditAsset>,
    source_amount: S,
) -> Result<PathsStrictSendRequest>

Creates a request to retrieve information about potential path payments.

The strict receive payment path endpoint lists the paths a payment can take based on the amount of an asset you want to send. The source asset amount stays constant, and the type and amount of an asset received varies based on offers in the order books.